All: Support RTL and LTR in RTL context
Supporting RTL in a LTR context is implicit
RTL is invoked by using a ui-rtl class wrapper with dir="rtl" attribute.
LTR in a RTL context is invoked by using a ui-ltr class wrapper and
accompanying dir="ltr" inside the ui-rtl wrapper.
ui-btn-start, ui-btn-end, ui-btn-icon-start and ui-btn-icon-end has been
added to allow easier ordering of elements,
as a replacement for *-left and *-right counterparts.
Tests have not been 100% updated with RTL considerations in this commit.
The only test updates are the ones needed to pass all current tests.
Transitions have not been included in RTL considerations.
RTL versions of demos have been copied and modified from
their LTR counterparts, and
can be accessed from the Right To Left option on the demos menu.

Ref gh-5167